services/auth-service/requirements.txt
# ./services/auth-service/requirements.txt
Flask==2.3.2
PyJWT==2.6.0 # For creating JWTs
redis==4.5.1 # Optional: If using Redis for sessions/blocklist
authlib==1.2.0
requests==2.31.0 # Needed by Authlib for fetching user info
python-dotenv==1.0.0